style(swiftui): swiftlint cleanup — session-introduced + mechanical violations (#1719)
