+++ fichero/fichero/Views/KnowledgeGraph/OntologyBrowser/ClaimSummaryCardView.swift
@@ -6,7 +6,7 @@ import SwiftUI
 
 struct ClaimSummaryCard: View {
     let claim: Components.Schemas.KnowledgeClaim
-    var onNavigateToSource: ((Components.Schemas.KnowledgeClaim) -> Void)? = nil
+    var onNavigateToSource: ((Components.Schemas.KnowledgeClaim) -> Void)?
 
     /// Expanded → reveals the verbatim source excerpt + fetches
     /// contradictions + evidence-chain. Collapsed by default to keep
